+++ test/test_data.rb
@@ -13,6 +13,8 @@ class TestData < Minitest::Test
 
     # Because some code is shared with Struct, check we don't share unnecessary functionality
     assert_raises(TypeError) { Data.define(:foo, keyword_init: true) }
+
+    refute_respond_to(Data.define, :define, "Cannot define from defined Data class")
   end
 
   def test_define_edge_cases
@@ -59,6 +61,7 @@ class TestData < Minitest::Test
     assert_equal(1, test.foo)
     assert_equal(2, test.bar)
     assert_equal(test, klass.new(1, 2))
+    assert_predicate(test, :frozen?)
 
     # Keywords
     test_kw = klass.new(foo: 1, bar: 2)
@@ -121,8 +124,6 @@ class TestData < Minitest::Test
     test = klass.new(bar: 2, foo: 1)
     assert_equal([1, 2], test.deconstruct)
 
-    assert_predicate(test, :frozen?)
-
     assert_kind_of(Integer, test.hash)
   end
 
@@ -165,6 +166,65 @@ class TestData < Minitest::Test
     refute_operator(o1, :eql?, o3)
   end
 
+  def test_with
+    klass = Data.define(:foo, :bar)
+    source = klass.new(foo: 1, bar: 2)
+
+    # Simple
+    test = source.with
+    assert_equal(source.object_id, test.object_id)
+
+    # Changes
+    test = source.with(foo: 10)
+
+    assert_equal(1, source.foo)
+    assert_equal(2, source.bar)
+    assert_equal(source, klass.new(foo: 1, bar: 2))
+
+    assert_equal(10, test.foo)
+    assert_equal(2, test.bar)
+    assert_equal(test, klass.new(foo: 10, bar: 2))
+
+    test = source.with(foo: 10, bar: 20)
+
+    assert_equal(1, source.foo)
+    assert_equal(2, source.bar)
+    assert_equal(source, klass.new(foo: 1, bar: 2))
+
+    assert_equal(10, test.foo)
+    assert_equal(20, test.bar)
+    assert_equal(test, klass.new(foo: 10, bar: 20))
+
+    # Keyword splat
+    changes = { foo: 10, bar: 20 }
+    test = source.with(**changes)
+
+    assert_equal(1, source.foo)
+    assert_equal(2, source.bar)
+    assert_equal(source, klass.new(foo: 1, bar: 2))
+
+    assert_equal(10, test.foo)
+    assert_equal(20, test.bar)
+    assert_equal(test, klass.new(foo: 10, bar: 20))
+
+    # Wrong protocol
+    assert_raises(ArgumentError, "wrong number of arguments (given 1, expected 0)") do
+      source.with(10)
+    end
+    assert_raises(ArgumentError, "unknown keywords: :baz, :quux") do
+      source.with(foo: 1, bar: 2, baz: 3, quux: 4)
+    end
+    assert_raises(ArgumentError, "wrong number of arguments (given 1, expected 0)") do
+      source.with(1, bar: 2)
+    end
+    assert_raises(ArgumentError, "wrong number of arguments (given 2, expected 0)") do
+      source.with(1, 2)
+    end
+    assert_raises(ArgumentError, "wrong number of arguments (given 1, expected 0)") do
+      source.with({ bar: 2 })
+    end
+  end
+
   def test_memberless
     klass = Data.define
 
@@ -185,6 +245,23 @@ class TestData < Minitest::Test
 
     assert_equal(10, distance.amount)
     assert_equal('km', distance.unit)
+  end
+
+  def test_dup
+    klass = Data.define(:foo, :bar)
+    test = klass.new(foo: 1, bar: 2)
+    assert_equal(klass.new(foo: 1, bar: 2), test.dup)
+    assert_predicate(test.dup, :frozen?)
+  end
+
+  Klass = Data.define(:foo, :bar)
+
+  def test_marshal
+    test = Klass.new(foo: 1, bar: 2)
+    loaded = Marshal.load(Marshal.dump(test))
+    assert_equal(test, loaded)
+    refute_same(test, loaded)
+    assert_predicate(loaded, :frozen?)
   end
 
   def test_namespaced_constant