+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- `dup` will return a frozen object
+- `define` cannot be used on Data subclasses
+- `with` returns `self` when given no arguments
+
 ## 1.0.3
 
 ### Fixed